% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autocorrelation.R
\name{acf_ly}
\alias{acf_ly}
\title{acf_ly
Auto Correlation Plot using R Plotly}
\usage{
acf_ly(series)
}
\arguments{
\item{series}{is numeric vector}
}
\value{
plot
}
\description{
acf_ly
Auto Correlation Plot using R Plotly
}
\examples{
series<- c(mtcars$mpg, mtcars$hp)
acf_ly(series)
}
