% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_Correlation.R
\name{ccf_ly}
\alias{ccf_ly}
\title{ccf_ly
Cross Correlation Plot using R Plotly}
\usage{
ccf_ly(series1 = NULL, series2 = NULL)
}
\arguments{
\item{series1}{is numeric vector}

\item{series2}{is numeric vector}
}
\value{
plot
}
\description{
ccf_ly
Cross Correlation Plot using R Plotly
}
\examples{
ccfq<- c(mtcars$mpg, mtcars$gear)
ccfa<- c(mtcars$hp, mtcars$cyl)
ccf_ly(series1=ccfq, series2=ccfa)
}
