% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef_pearson.R
\name{corr_coef_pearson}
\alias{corr_coef_pearson}
\title{corr_coef_pearson
correlation Coeficient using pearson Method}
\usage{
corr_coef_pearson(variable1 = NULL, variable2 = NULL, decimal = NULL)
}
\arguments{
\item{variable1}{is a column or Numeric vector}

\item{variable2}{is a column or Numeric vector}

\item{decimal}{is a Numeric Argument}
}
\value{
correlation coeficient
}
\description{
corr_coef_pearson
correlation Coeficient using pearson Method
}
\examples{
corr_coef_pearson(variable1= mtcars$disp, variable2=mtcars$hp, decimal = 2)
}
