% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef_spearman.R
\name{corr_coef_spearman}
\alias{corr_coef_spearman}
\title{corr_coef_spearman
correlation Coeficient using spearman Method}
\usage{
corr_coef_spearman(variable1 = NULL, variable2 = NULL, decimal = NULL)
}
\arguments{
\item{variable1}{is a column or Numeric vector}

\item{variable2}{is a column or Numeric vector}

\item{decimal}{is a Numeric Argument}
}
\value{
correlation coeficient
}
\description{
corr_coef_spearman
correlation Coeficient using spearman Method
}
\examples{
corr_coef_spearman(variable1= cars$speed, variable2=cars$dist, decimal = 2)
}
