% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_scatterly.R
\name{corr_scatterly}
\alias{corr_scatterly}
\title{corr_scatterly
Correlation Scatter Plot using R Plotly}
\usage{
corr_scatterly(data = NULL, x = NULL, y = NULL, corr_coef = NULL,
  xname = "", yname = "")
}
\arguments{
\item{data}{is a Data Frame}

\item{x}{is a column of Data Frame}

\item{y}{is a column of Data Frame}

\item{corr_coef}{is a Correlation Coeficient}

\item{xname}{is a string and Name of X-axis}

\item{yname}{is a string and Name of Y-axis}
}
\value{
plot
}
\description{
corr_scatterly
Correlation Scatter Plot using R Plotly
}
\examples{
spearman<- corr_coef_spearman(variable1= cars$speed, variable2=cars$dist, decimal = 2)
corr_scatterly(data=cars,x=cars$speed,y=cars$dist,corr_coef=spearman,xname="speed",yname="dist")
}
