% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_function.R
\name{grouping_function}
\alias{grouping_function}
\title{grouping_function}
\usage{
grouping_function(parameter_value = NULL, subgroup_size = NULL)
}
\arguments{
\item{parameter_value}{is a column of Data Frame}

\item{subgroup_size}{is group size of vector}
}
\value{
grouped data frame
}
\description{
grouping_function
}
\examples{
grouping_function(parameter_value = cars$speed, subgroup_size = 2)
}
