% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-Chart.R
\name{ichart_ly}
\alias{ichart_ly}
\title{ichart_ly
I Chart or X-Chart Plot using R Plotly}
\usage{
ichart_ly(data, x = NULL, y = NULL, xname = NULL, yname = NULL)
}
\arguments{
\item{data}{is a Data Frame}

\item{x}{is a column of data frame on x-axis}

\item{y}{is a column of data frame on y-axis}

\item{xname}{is a string and Name of x-axis}

\item{yname}{is a string and Name of y-axis}
}
\value{
plot
}
\description{
ichart_ly
I Chart or X-Chart Plot using R Plotly
}
\examples{
ichart_ly(mtcars, x = mtcars$mpg, y = mtcars$hp, xname = "mpg", yname = "hp")
}
