% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixly.R
\name{matrixly}
\alias{matrixly}
\title{matrixly
Correlation matrix Plot using R Plotly}
\usage{
matrixly(data = NULL)
}
\arguments{
\item{data}{is a Data Frame}
}
\value{
plot
}
\description{
matrixly
Correlation matrix Plot using R Plotly
}
\examples{
matrixly(data = mtcars)
}
