% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal_Distribution.R
\name{normal_distribution}
\alias{normal_distribution}
\title{normal_distribution
normal distribution Chart using R Plotly}
\usage{
normal_distribution(data = NULL, parameter = NULL, xname = "")
}
\arguments{
\item{data}{is a Data Frame}

\item{parameter}{is a column of data frame}

\item{xname}{is title of xaxis}
}
\value{
plot
}
\description{
normal_distribution
normal distribution Chart using R Plotly
}
\examples{
normal_distribution(data = mtcars, parameter = mtcars$mpg, xname = "mpg")
}
